Add cbuff_uint64 variant; fix %d -> %llu in inspect function
- generate.sh: add uint64_t/uint64 variant (one line)
- cbuff_uint64.h, cbuff_uint64_test.c: generated, all 10 tests pass
- cbuff_uint8_test.c: fix circularBuffer_inspect to use %llu with cast
  so the format string is correct for all types including uint64
- Makefile, clib.json, .gitignore, validate.sh: updated for uint64

Note: uint64 also covers double by the same casting convention
(buffer is bit-agnostic, caller interprets the bits).

+++ cbuff_uint16_test.c
@@ -19,7 +19,7 @@ void circularBuffer_inspect(cbuff_uint16_t *cb, char *msg)
     printf("\n%s : ", msg);
     for (int i = 0; i < cb->capacity; i++)
     {
-        printf(" %d", cb->buffer[i]);
+        printf(" %llu", (unsigned long long)cb->buffer[i]);
     }
     printf("\n");
 }
